XD-530 Add spring-xd-reactor Project
Added reactor-syslog XD module.

Changed to eventLoop dispatcher.

XD-530 Adds support for Reactor via two new abstractions: ReactorSubscribableChannel and SyslogInboundChannelAdapter. The former is backed by a Reactor Processor, which has higher overall throughput with the trade-off that each new instance creates a new Thread of execution for the underlying Disruptor RingBuffer. The latter uses the Reactor TcpServer with the built-in syslog Codec to listen on a port for syslog messages, which it simply passes into the output channel without modification. Included in this commit is a basic XML schema for a Reactor namespace and a couple of integration tests.

XD-530 Rename channel implementation to ReactorPublishSubscribeChannel and update the schema to reflect the behavior of this implementation. Also added the submodule to the project.

XD-530 Rename channel implementation to ReactorPublishSubscribeChannel and update the schema to reflect the behavior of this implementation. Also added the submodule to the project.

XD-530 Added project definition to the build.gradle.

XD-530 Fix misspelling

XD-530 Changed host and port configuration to properties.

XD-530 Suppress compiler warnings.

XD-530 Change syslog config to accept host and port via properties, not just a port number via CTOR.

XD-530 Removed the getLatch() method as it was confusing the ServiceActivator invoker and causing intermittent test failures.

XD-530 Updated copyright. Added spring.tooling file. Tweaked a variable name.